Added Comments and Internal Documentation
- Internal Documentation Update
- Better Foldering of Pages
