+++ medicrypt/src/components/sections/AnalyticsSelect.js
@@ -1,6 +1,47 @@
 import React from 'react';
 import { IoMdArrowDropdown } from "react-icons/io";
 
+/**
+ * AnalyticsSelect Component
+ *
+ * The `AnalyticsSelect` component provides a dropdown interface for selecting a video from a 
+ * list of available options. It enhances user experience by allowing researchers to easily choose 
+ * a video for which they want to view metrics. The component is styled to blend seamlessly with 
+ * other UI elements while providing a clear and functional design.
+ *
+ * Props:
+ * -------
+ * @param {string} value - The currently selected value of the dropdown, representing the 
+ *                         selected video.
+ * @param {function} onChange - A callback function to handle changes in selection. This function 
+ *                              is triggered when the user selects a different option from the dropdown.
+ * @param {string} className - Custom CSS class names for additional styling of the dropdown wrapper.
+ * @param {React.ReactNode} children - The options to be displayed within the dropdown, typically 
+ *                                      consisting of <option> elements.
+ *
+ * Usage:
+ * ------
+ * The `AnalyticsSelect` component is intended for use in analytics dashboards where users need 
+ * to select specific videos to analyze metrics related to those videos.
+ *
+ * Example:
+ * -------
+ * <AnalyticsSelect value={selectedVideo} onChange={handleVideoChange} className="my-select">
+ *   <option value="video1.mp4">Video 1</option>
+ *   <option value="video2.mp4">Video 2</option>
+ * </AnalyticsSelect>
+ *
+ * Dependencies:
+ * -------------
+ * - React: Core library for component rendering.
+ * - IoMdArrowDropdown: An icon from the `react-icons` library to visually indicate the dropdown 
+ *                      functionality.
+ *
+ * Code Author:
+ * ------------
+ * - Charles Andre C. Bandala
+ */
+
 const AnalyticsSelect = ({ value, onChange, className, children }) => {
   return (
     <div className={`relative ${className}`}>
