+++ medicrypt/src/pages-css/General.css
@@ -1,4 +1,3 @@
 input[type=text], input[type=password]{
-    
     color: '#21316F';
 }
