+++ medicrypt/src/pages/MenuPages/Explore.js
