+++ medicrypt/src/pages/MenuPages/HowToUse.js
